requirements.txt
openai
pyautogui
pillow
opencv-python
numpy
requests
mcp[cli]